Collect the results from a parametric bootstrap MGPM fit

Collect the results from a parametric bootstrap MGPM fit

CollectBootstrapResults(resultDir, prefixFiles, ids, inferredTree,
  epochs = seq(minLength, max(PCMTreeNodeTimes(inferredTree)), by = 10 *
  minLength), minLength = 0.2,
  exprResultFile = "paste0(prefixFiles, id, \"/Result_\", prefixFiles, id, \".RData\")",
  nameFitObject = "fitMappings", keepBootstrapData = FALSE,
  verbose = FALSE)

Arguments

resultDir, prefixFiles
character strings denoting the directory and the file prefix used to locate the result .RData files. The way how the result file-names are formed is specified by the argument exprResultFile.

ids
an integer vector denoting the ids of the bootstrap iterations.

inferredTree
a PCMTree or a phylo object with set partition as inferred during the ML or recursive clade partition search. This is the tree used to simulate the bootstrap datasets.

epochs
a double vector denoting in increasing order the time-points at which ancestral mean and regression slope parameters should be calculated. The inferredTree is to be 'sliced' by inserting singleton nodes at all branch crossings with each epoch. By default the epochs are set at equal intervals of length 10*minLength, via the following code: seq(minLength, max(PCMTreeNodeTimes(inferredTree)), by = 10*minLength). See also minLength.

minLength
a positive double denoting the minimal time distance between a node in the tree and an epoch in epochs, at which an additional singleton node can be inserted. Default: 0.2.

exprResultFile
a character string evaluating to an R expression. Default setting: 'paste0(prefixFiles, id, "/Result_", prefixFiles, id, ".RData")' In the above R-expression, id is any integer element from ids.

nameFitObject
a character string (default: 'fitMappings') denoting the name of the PCMFitModelMappings object inside each result .RData file (see also resultDir and prefixFiles).

keepBootstrapData
logical indicating if the returned object should contain the bootstrap simulated trait data.

verbose
logical indicating if some debug or information messages should be printed during the collection of the results.

Value

a named list of S3 class 'ParametricBootstrapFits'.