Optimal information score search for a mixed Gaussian phylogenetic model, given a tree, trait measurements at its tips and a score function.

A mixed Gaussian phylogenetic model (MGPM) represents a Gaussian phylogenetic model with shifts in the underlying parameters and, optionally, type of Gaussian stochastic process (e.g. shifts from a BM to an OU model of evolution). Formally, an MGPM consists of the following components:

A shift-point configuration: this is a subset of the nodes in the tree including at least the root node

b

PCMFitMixed(X, tree, modelTypes = MGPMDefaultModelTypes(),
  subModels = c(B = "A", C = "A", D = "B", E = "D", F = "E"),
  argsMixedGaussian = Args_MixedGaussian_MGPMDefaultModelTypes(),
  SE = matrix(0, nrow(X), PCMTreeNumTips(tree)),
  generatePCMModelsFun = PCMGenerateModelTypes, metaIFun = PCMInfo,
  positiveValueGuard = Inf, scoreFun = AIC, fitMappingsPrev = NULL,
  tableFitsPrev = fitMappingsPrev$tableFits,
  modelTypesInTableFitsPrev = NULL, listPartitions = NULL,
  minCladeSizes = 20L, skipNodes = character(),
  maxCladePartitionLevel = if (is.null(listPartitions)) 8L else 1L,
  maxNumNodesPerCladePartition = 1L,
  listAllowedModelTypesIndices = c("best-clade-2", "best-clade", "all"),
  argsConfigOptim1 = DefaultArgsConfigOptim(numCallsOptim = 10),
  argsConfigOptim2 = DefaultArgsConfigOptim(numCallsOptim = 4),
  argsConfigOptim3 = DefaultArgsConfigOptim(numCallsOptim = 10),
  maxNumRoundRobins = 0, maxNumPartitionsInRoundRobins = 2,
  listPCMOptions = PCMOptions(), skipFitWhenFoundInTableFits = TRUE,
  doParallel = FALSE, prefixFiles = "fits_",
  saveTempWorkerResults = TRUE, printFitVectorsToConsole = FALSE,
  verbose = TRUE, debug = FALSE)

Arguments

X
a k x N numerical matrix with possible NA and NaN entries. Each column of X contains the measured trait values for one species (tip in tree). Missing values can be either not-available (NA) or not existing (NaN). Thse two values have are treated differently when calculating likelihoods: see PCMPresentCoordinates.

tree
a phylo object with N tips.

SE
a k x N matrix specifying the standard error for each measurement in X. Alternatively, a k x k x N cube specifying an upper triangular k x k Cholesky factor of the variance covariance matrix for the measurement error for each node i=1, ..., N. Default: matrix(0.0, PCMNumTraits(model), PCMTreeNumTips(tree)).

positiveValueGuard
a real number (not necessarily positive) used during the fit as a threshold for highly positive but likely incorrect log-likelihood values. This argument is set to Inf by default and should be used only as a last escape. The recommended way to of preventing such numerical errors is by setting more stringent values for the runtime options PCMBase.Threshold.EV and PCMBase.Threshold.SV (see PCMOptions).

doParallel
logical indicating if optim calls should be executed in parallel using the foreach() %dopar% {} construct. Default: FALSE.

verbose
logical indicating if information messages should be printed to the console while running. Default: FALSE.

Value

an S3 object of class PCMFitModelMappings.